Moved all functions from weather/tasks.py into weather/utils.py.
